requirements.txt
numpy
pandas 
seaborn
matplotlib
scikit-learn

python-dotenv
pymysql
mysql-connector-python
requests